refactor: 전체적인 exception 패키지 수정 - 도메인 구조 패키지 방식으로 exception 패키지 수정 - final, abstract로 하위 클래스와 상위 클래스 구분
