+++ src/main/java/wooteco/subway/exception/BadRequestException.java
@@ -1,10 +1,9 @@
-package wooteco.subway.exception.badrequest;
+package wooteco.subway.exception;
 
 import org.springframework.http.HttpStatus;
-import wooteco.subway.exception.SubwayException;
 import wooteco.subway.exception.dto.ErrorResponse;
 
-public class BadRequestException extends SubwayException {
+public abstract class BadRequestException extends SubwayException {
     public BadRequestException(String message) {
         super(HttpStatus.BAD_REQUEST, new ErrorResponse(message));
     }
