+++ src/main/java/wooteco/subway/exception/InvalidDistanceException.java
@@ -1,6 +1,6 @@
-package wooteco.subway.exception.notfound;
+package wooteco.subway.exception;
 
-public class InvalidDistanceException extends NotFoundException {
+public final class InvalidDistanceException extends NotFoundException {
     public InvalidDistanceException() {
         super("계산 가능한 거리를 초과하였습니다.");
     }
