+++ src/main/java/wooteco/subway/exception/NotFoundException.java
@@ -1,10 +1,9 @@
-package wooteco.subway.exception.notfound;
+package wooteco.subway.exception;
 
 import org.springframework.http.HttpStatus;
-import wooteco.subway.exception.SubwayException;
 import wooteco.subway.exception.dto.ErrorResponse;
 
-public class NotFoundException extends SubwayException {
+public abstract class NotFoundException extends SubwayException {
     public NotFoundException(String message) {
         super(HttpStatus.NOT_FOUND, new ErrorResponse(message));
     }
