+++ src/main/java/wooteco/subway/exception/UnauthorizedException.java
@@ -1,10 +1,9 @@
-package wooteco.subway.exception.unauthorized;
+package wooteco.subway.exception;
 
 import org.springframework.http.HttpStatus;
-import wooteco.subway.exception.SubwayException;
 import wooteco.subway.exception.dto.ErrorResponse;
 
-public class UnauthorizedException extends SubwayException {
+public abstract class UnauthorizedException extends SubwayException {
     public UnauthorizedException(String message) {
         super(HttpStatus.UNAUTHORIZED, new ErrorResponse(message));
     }
