+++ src/main/java/wooteco/subway/exception/UnprocessableEntityException.java
@@ -1,10 +1,9 @@
-package wooteco.subway.exception.unprocessableentity;
+package wooteco.subway.exception;
 
 import org.springframework.http.HttpStatus;
-import wooteco.subway.exception.SubwayException;
 import wooteco.subway.exception.dto.ErrorResponse;
 
-public class UnprocessableEntityException extends SubwayException {
+public abstract class UnprocessableEntityException extends SubwayException {
     public UnprocessableEntityException(String message) {
         super(HttpStatus.UNPROCESSABLE_ENTITY, new ErrorResponse(message));
     }
