--- src/main/java/wooteco/subway/exception/notfound/NotFoundStationException.java
@@ -1,7 +1,0 @@
-package wooteco.subway.exception.notfound;
-
-public class NotFoundStationException extends NotFoundException {
-    public NotFoundStationException() {
-        super("존재하지 않는 지하철 역입니다.");
-    }
-}
