--- src/main/java/wooteco/subway/exception/notfound/NotFoundStationOfSectionException.java
@@ -1,7 +1,0 @@
-package wooteco.subway.exception.notfound;
-
-public class NotFoundStationOfSectionException extends NotFoundException {
-    public NotFoundStationOfSectionException() {
-        super("상행 종점역이 존재하지 않습니다.");
-    }
-}
