+++ src/main/java/wooteco/subway/line/exception/DuplicateLineException.java
@@ -1,0 +1,9 @@
+package wooteco.subway.line.exception;
+
+import wooteco.subway.exception.BadRequestException;
+
+public final class DuplicateLineException extends BadRequestException {
+    public DuplicateLineException() {
+        super("중복된 지하철 노선입니다.");
+    }
+}
