+++ src/main/java/wooteco/subway/line/exception/ExistStationInSectionException.java
@@ -1,0 +1,9 @@
+package wooteco.subway.line.exception;
+
+import wooteco.subway.exception.BadRequestException;
+
+public final class ExistStationInSectionException extends BadRequestException {
+    public ExistStationInSectionException() {
+        super("구간에 존재하는 지하철 역을 삭제할 수 없습니다.");
+    }
+}
