+++ src/main/java/wooteco/subway/member/application/MemberService.java
@@ -2,13 +2,13 @@ package wooteco.subway.member.application;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import wooteco.subway.exception.notfound.NoMemberException;
-import wooteco.subway.exception.unprocessableentity.DuplicateEmailException;
 import wooteco.subway.member.dao.MemberDao;
 import wooteco.subway.member.domain.Member;
 import wooteco.subway.member.domain.User;
 import wooteco.subway.member.dto.MemberRequest;
 import wooteco.subway.member.dto.MemberResponse;
+import wooteco.subway.member.exception.DuplicateEmailException;
+import wooteco.subway.member.exception.NoMemberException;
 
 @Service
 public class MemberService {
