+++ src/main/java/wooteco/subway/member/domain/Member.java
@@ -1,7 +1,7 @@
 package wooteco.subway.member.domain;
 
 import org.apache.commons.lang3.StringUtils;
-import wooteco.subway.exception.unauthorized.AuthorizationException;
+import wooteco.subway.auth.exception.AuthorizationException;
 
 public class Member {
     private Long id;
