+++ src/main/java/wooteco/subway/member/domain/User.java
@@ -4,8 +4,12 @@ import wooteco.subway.line.domain.fare.AgeFarePolicy;
 
 public interface User {
     Long getId();
+
     String getEmail();
+
     Integer getAge();
+
     AgeFarePolicy farePolicy();
+
     boolean isOverThan(int age);
 }
