+++ src/main/java/wooteco/subway/member/exception/DuplicateEmailException.java
@@ -1,0 +1,9 @@
+package wooteco.subway.member.exception;
+
+import wooteco.subway.exception.UnprocessableEntityException;
+
+public final class DuplicateEmailException extends UnprocessableEntityException {
+    public DuplicateEmailException() {
+        super("이메일이 중복되었습니다.");
+    }
+}
