+++ src/main/java/wooteco/subway/path/application/PathFinder.java
@@ -3,11 +3,11 @@ package wooteco.subway.path.application;
 import org.jgrapht.GraphPath;
 import org.jgrapht.alg.shortestpath.DijkstraShortestPath;
 import org.springframework.stereotype.Service;
-import wooteco.subway.exception.notfound.InvalidPathException;
 import wooteco.subway.line.domain.Line;
 import wooteco.subway.path.domain.SectionEdge;
 import wooteco.subway.path.domain.SubwayGraph;
 import wooteco.subway.path.domain.SubwayPath;
+import wooteco.subway.path.exception.InvalidPathException;
 import wooteco.subway.station.domain.Station;
 
 import java.util.List;
