+++ src/main/java/wooteco/subway/station/exception/DuplicateStationException.java
@@ -1,0 +1,9 @@
+package wooteco.subway.station.exception;
+
+import wooteco.subway.exception.BadRequestException;
+
+public final class DuplicateStationException extends BadRequestException {
+    public DuplicateStationException() {
+        super("이미 존재하는 지하철 역입니다.");
+    }
+}
