+++ src/main/java/wooteco/subway/station/exception/NotFoundStationException.java
@@ -1,0 +1,9 @@
+package wooteco.subway.station.exception;
+
+import wooteco.subway.exception.NotFoundException;
+
+public final class NotFoundStationException extends NotFoundException {
+    public NotFoundStationException() {
+        super("존재하지 않는 지하철 역입니다.");
+    }
+}
